refactor: order controller
