+++ src/controllers/orders.controllers.js
@@ -1,1463 +1,2 @@
-import mongoose from "mongoose";
-import { asyncHandler } from "../utils/asyncHandler.js";
-import { ApiError } from "../utils/ApiError.js";
-import { ApiResponse } from "../utils/ApiResponse.js";
-import Order from "../models/order.models.js";
-import EscrowWallet from "../models/escrowWallet.models.js";
-import Notification from "../models/notification.models.js";
-import Message from "../models/message.models.js";
-import Chat from "../models/chat.models.js";
-import socketManager from "../config/socket.js";
-import notificationCache from "../utils/notificationCache.utils.js";
-
-const safeEmitToChat = (chatId, event, data) => {
-    if (socketManager.isReady()) {
-        socketManager.emitToChat(chatId, event, data);
-    }
-};
-
-const safeEmitToUser = (userId, event, data) => {
-    if (socketManager.isReady()) {
-        socketManager.emitToUser(userId, event, data);
-    }
-};
-
-const sendOrderNotification = async ({
-    recipientId,
-    senderId,
-    orderId,
-    orderNumber,
-    notificationMessage,
-    chatMessageText,
-    chatId,
-    buyerId
-}) => {
-    try {
-        // Create database notification
-        const notification = await Notification.create({
-            receiverId: recipientId,
-            senderId: senderId,
-            type: 'order',
-            orderId: orderId,
-            message: notificationMessage
-        });
-
-        if (global.io) {
-            global.io.to(`user_${recipientId}`).emit("notification", notification);
-        }
-
-        await notificationCache.invalidateNotificationCache(recipientId);
-
-        // Send automated chat message (skip for guest/shareable orders)
-        if (!chatId || !buyerId) return;
-
-        const chat = await Chat.findById(chatId);
-        if (!chat) return;
-
-        const recipients = chat.participants.filter(
-            p => p.toString() !== senderId.toString()
-        );
-
-        const chatMessage = await Message.create({
-            chatId,
-            sender: senderId,
-            message: chatMessageText,
-            messageType: 'order_update',
-            timestamp: new Date(),
-            readBy: [senderId],
-            deliveryStatus: recipients.map(rid => ({
-                userId: rid,
-                status: 'sent'
-            }))
-        });
-
-        chat.lastMessageAt = new Date();
-        chat.lastMessage = {
-            sender: senderId,
-            message: chatMessageText,
-            timestamp: new Date()
-        };
-        chat.lastMessageId = chatMessage._id;
-        await chat.save();
-
-        const populatedMessage = await Message.findById(chatMessage._id)
-            .populate('sender', 'username fullName profileImageUrl')
-            .lean();
-
-        safeEmitToChat(chatId.toString(), 'new_message', {
-            chatId,
-            message: populatedMessage
-        });
-
-        safeEmitToUser(recipientId.toString(), 'new_message', {
-            chatId,
-            message: populatedMessage
-        });
-
-        await notificationCache.invalidateMessageCache(recipientId);
-    } catch (error) {
-        console.error(`sendOrderNotification error for order ${orderNumber}:`, error);
-    }
-};
-
-// Get order details
-export const getOrderDetails = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const userId = req.user._id;
-
-    const order = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl')
-        .populate('sellerId', 'fullName username profileImageUrl')
-        .populate('postId', 'media caption');
-
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.buyerId._id.toString() !== userId.toString() &&
-        order.sellerId._id.toString() !== userId.toString()) {
-        throw new ApiError(403, "Not authorized to view this order");
-    }
-
-    const responseData = { order };
-
-    // Include dispute policy info when order is disputed
-    if (order.orderStatus === 'disputed' && order.dispute) {
-        responseData.disputePolicy = {
-            videoRequired: true,
-            videoUploaded: !!order.dispute.disputeVideoUrl,
-            message: "The buyer must upload a video showing proof of damage to be eligible for refund or return. This footage is accessible to both the admin and the seller. If valid proof is not provided, the item will not be eligible for refund or return."
-        };
-    }
-
-    return res.status(200).json(
-        new ApiResponse(200, responseData, "Order details fetched")
-    );
-});
-
-// Get buyer's orders
-export const getBuyerOrders = asyncHandler(async (req, res) => {
-    const { status, page = 1, limit = 20 } = req.query;
-    const buyerId = req.user._id;
-
-    const query = { buyerId };
-    if (status) query.orderStatus = status;
-
-    const orders = await Order.find(query)
-        .populate('sellerId', 'fullName username profileImageUrl')
-        .sort({ createdAt: -1 })
-        .skip((page - 1) * limit)
-        .limit(parseInt(limit));
-
-    const total = await Order.countDocuments(query);
-
-    return res.status(200).json(
-        new ApiResponse(200, { orders, total, page: parseInt(page), totalPages: Math.ceil(total / limit) }, "Buyer orders fetched")
-    );
-});
-
-// Get count of new orders for seller (orders that need to be shipped)
-export const getSellerNewOrdersCount = asyncHandler(async (req, res) => {
-    const sellerId = req.user._id;
-
-    // Count orders with 'payment_received' status - these need seller action (shipping)
-    const newOrdersCount = await Order.countDocuments({
-        sellerId,
-        orderStatus: 'payment_received'
-    });
-
-    return res.status(200).json(
-        new ApiResponse(200, { newOrdersCount }, "New orders count fetched")
-    );
-});
-
-// Get seller's orders
-export const getSellerOrders = asyncHandler(async (req, res) => {
-    const { status, page = 1, limit = 20 } = req.query;
-    const sellerId = req.user._id;
-
-    const query = { sellerId };
-    if (status) query.orderStatus = status;
-
-    const orders = await Order.find(query)
-        .populate('buyerId', 'fullName username profileImageUrl')
-        .sort({ createdAt: -1 })
-        .skip((page - 1) * limit)
-        .limit(parseInt(limit));
-
-    const total = await Order.countDocuments(query);
-
-    return res.status(200).json(
-        new ApiResponse(200, { orders, total, page: parseInt(page), totalPages: Math.ceil(total / limit) }, "Seller orders fetched")
-    );
-});
-
-// Seller confirms the placed order
-export const sellerConfirmOrder = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const sellerId = req.user._id;
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.sellerId.toString() !== sellerId.toString()) {
-        throw new ApiError(403, "Only the seller can confirm this order");
-    }
-
-    if (order.orderStatus !== 'payment_received') {
-        throw new ApiError(400, "Order can only be confirmed when payment is received");
-    }
-
-    order.orderStatus = 'processing';
-    order.sellerResponse = {
-        status: 'confirmed',
-        respondedAt: new Date()
-    };
-    await order.save();
-
-    // Notify buyer that seller has confirmed the order
-    if (order.buyerId) {
-        sendOrderNotification({
-            recipientId: order.buyerId,
-            senderId: sellerId,
-            orderId: order._id,
-            orderNumber: order.orderNumber,
-            notificationMessage: `Your order #${order.orderNumber} has been confirmed by the seller! It is now being processed.`,
-            chatMessageText: `Order #${order.orderNumber} has been confirmed by the seller!\n\nYour order is now being processed and will be shipped soon.`,
-            chatId: order.chatId,
-            buyerId: order.buyerId
-        }).catch(err => console.error('Seller confirm notification error:', err));
-    }
-
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, { order: updatedOrder }, "Order confirmed by seller")
-    );
-});
-
-// Seller rejects the placed order with reason
-export const sellerRejectOrder = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const { reason, note } = req.body;
-    const sellerId = req.user._id;
-
-    if (!reason) {
-        throw new ApiError(400, "Rejection reason is required");
-    }
-
-    const validReasons = ['out_of_stock', 'price_change', 'invalid_address', 'need_clarification', 'certificate_required', 'other'];
-    if (!validReasons.includes(reason)) {
-        throw new ApiError(400, `Invalid rejection reason. Must be one of: ${validReasons.join(', ')}`);
-    }
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.sellerId.toString() !== sellerId.toString()) {
-        throw new ApiError(403, "Only the seller can reject this order");
-    }
-
-    if (order.orderStatus !== 'payment_received') {
-        throw new ApiError(400, "Order can only be rejected when payment is received");
-    }
-
-    const reasonLabels = {
-        out_of_stock: 'Out of Stock',
-        price_change: 'Price Change',
-        invalid_address: 'Buyer Address Not Valid',
-        need_clarification: 'Seller Needs More Clarification',
-        certificate_required: 'Certificates Required for This Product',
-        other: 'Other'
-    };
-
-    order.orderStatus = 'seller_rejected';
-    order.sellerResponse = {
-        status: 'rejected',
-        rejectionReason: reason,
-        rejectionNote: note || '',
-        respondedAt: new Date()
-    };
-    await order.save();
-
-    // Notify buyer that seller has rejected the order
-    const rejectionLabel = reasonLabels[reason] || reason;
-    if (order.buyerId) {
-        sendOrderNotification({
-            recipientId: order.buyerId,
-            senderId: sellerId,
-            orderId: order._id,
-            orderNumber: order.orderNumber,
-            notificationMessage: `Your order #${order.orderNumber} has been rejected by the seller. Reason: ${rejectionLabel}`,
-            chatMessageText: `Order #${order.orderNumber} has been rejected by the seller.\n\nReason: ${rejectionLabel}${note ? `\nNote: ${note}` : ''}\n\nYour payment is safe in escrow. The Findernate team will process your refund shortly.`,
-            chatId: order.chatId,
-            buyerId: order.buyerId
-        }).catch(err => console.error('Seller reject notification error:', err));
-    }
-
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, { order: updatedOrder }, "Order rejected by seller. Admin will process the refund.")
-    );
-});
-
-// Seller marks order as shipped
-export const markOrderShipped = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const { trackingId, carrier, packingVideoUrl, packingImages } = req.body;
-    const sellerId = req.user._id;
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.sellerId.toString() !== sellerId.toString()) {
-        throw new ApiError(403, "Only the seller can mark order as shipped");
-    }
-
-    if (order.orderStatus !== 'payment_received' && order.orderStatus !== 'processing') {
-        throw new ApiError(400, "Order cannot be shipped in current status");
-    }
-
-    order.orderStatus = 'shipped';
-    order.shippingInfo = {
-        ...order.shippingInfo,
-        trackingId: trackingId || order.shippingInfo?.trackingId,
-        carrier: carrier || order.shippingInfo?.carrier,
-        shippedAt: new Date(),
-        packingVideoUrl: packingVideoUrl || order.shippingInfo?.packingVideoUrl,
-        packingImages: packingImages && packingImages.length > 0
-            ? [...(order.shippingInfo?.packingImages || []), ...packingImages].slice(0, 10)
-            : order.shippingInfo?.packingImages
-    };
-    await order.save();
-
-    // Notify buyer that order has been shipped
-    if (order.buyerId) {
-        sendOrderNotification({
-            recipientId: order.buyerId,
-            senderId: sellerId,
-            orderId: order._id,
-            orderNumber: order.orderNumber,
-            notificationMessage: `Your order #${order.orderNumber} has been shipped!`,
-            chatMessageText: `Order #${order.orderNumber} has been shipped!${order.shippingInfo?.trackingId ? `\nTracking ID: ${order.shippingInfo.trackingId}` : ''}${order.shippingInfo?.carrier ? `\nCarrier: ${order.shippingInfo.carrier}` : ''}\n\nYou will be notified when it is delivered.`,
-            chatId: order.chatId,
-            buyerId: order.buyerId
-        }).catch(err => console.error('Ship notification error:', err));
-    }
-
-    // Populate and return updated order
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, { order: updatedOrder }, "Order marked as shipped")
-    );
-});
-
-// Seller updates tracking info (tracking ID, carrier) after shipping
-export const updateTrackingInfo = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const { trackingId, carrier } = req.body;
-    const sellerId = req.user._id;
-
-    if (!trackingId && !carrier) {
-        throw new ApiError(400, "Provide at least trackingId or carrier to update");
-    }
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.sellerId.toString() !== sellerId.toString()) {
-        throw new ApiError(403, "Only the seller can update tracking info");
-    }
-
-    if (order.orderStatus !== 'shipped' && order.orderStatus !== 'delivered') {
-        throw new ApiError(400, "Order must be shipped or delivered to update tracking info");
-    }
-
-    if (!order.shippingInfo) {
-        order.shippingInfo = {};
-    }
-    if (trackingId) order.shippingInfo.trackingId = trackingId;
-    if (carrier) order.shippingInfo.carrier = carrier;
-    await order.save();
-
-    // Notify buyer about updated tracking info
-    if (order.buyerId) {
-        sendOrderNotification({
-            recipientId: order.buyerId,
-            senderId: sellerId,
-            orderId: order._id,
-            orderNumber: order.orderNumber,
-            notificationMessage: `Tracking info updated for order #${order.orderNumber}`,
-            chatMessageText: `Tracking info updated for Order #${order.orderNumber}${trackingId ? `\nTracking ID: ${trackingId}` : ''}${carrier ? `\nCarrier: ${carrier}` : ''}`,
-            chatId: order.chatId,
-            buyerId: order.buyerId
-        }).catch(err => console.error('Tracking update notification error:', err));
-    }
-
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, { order: updatedOrder }, "Tracking info updated")
-    );
-});
-
-// Seller marks order as delivered
-export const markOrderDelivered = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const sellerId = req.user._id;
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.sellerId.toString() !== sellerId.toString()) {
-        throw new ApiError(403, "Only the seller can mark order as delivered");
-    }
-
-    if (order.orderStatus !== 'shipped') {
-        throw new ApiError(400, "Order must be shipped first");
-    }
-
-    order.orderStatus = 'delivered';
-    if (!order.shippingInfo) {
-        order.shippingInfo = {};
-    }
-    order.shippingInfo.deliveredAt = new Date();
-    await order.save();
-
-    // Notify buyer that order has been delivered
-    if (order.buyerId) {
-        sendOrderNotification({
-            recipientId: order.buyerId,
-            senderId: sellerId,
-            orderId: order._id,
-            orderNumber: order.orderNumber,
-            notificationMessage: `Your order #${order.orderNumber} has been marked as delivered`,
-            chatMessageText: `Order #${order.orderNumber} has been marked as delivered by the seller.\n\nPlease confirm delivery once you have received and inspected the product.`,
-            chatId: order.chatId,
-            buyerId: order.buyerId
-        }).catch(err => console.error('Deliver notification error:', err));
-    }
-
-    // Populate and return updated order
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, { order: updatedOrder }, "Order marked as delivered")
-    );
-});
-
-// Buyer confirms delivery - releases payment to seller
-export const confirmDelivery = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const { rating, review, openingVideoUrl } = req.body;
-    const buyerId = req.user._id;
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.buyerId.toString() !== buyerId.toString()) {
-        throw new ApiError(403, "Only the buyer can confirm delivery");
-    }
-
-    if (order.orderStatus !== 'delivered' && order.orderStatus !== 'shipped') {
-        throw new ApiError(400, "Order must be shipped or delivered first");
-    }
-
-    order.orderStatus = 'confirmed';
-    // Payment stays held — admin will manually release
-    order.deliveryConfirmedAt = new Date();
-
-    if (rating) order.buyerRating = rating;
-    if (review) order.buyerReview = review;
-    if (openingVideoUrl) {
-        order.buyerProof = { ...order.buyerProof, openingVideoUrl, uploadedAt: new Date() };
-    }
-
-    await order.save();
-
-    // Notify seller that delivery has been confirmed by the buyer
-    sendOrderNotification({
-        recipientId: order.sellerId,
-        senderId: buyerId,
-        orderId: order._id,
-        orderNumber: order.orderNumber,
-        notificationMessage: `Order #${order.orderNumber} delivery confirmed by buyer! Payment will be released by admin shortly.`,
-        chatMessageText: `Order #${order.orderNumber} has been delivered successfully!\n\nThe buyer has confirmed receiving the product. Payment will be released by admin shortly.${rating ? `\nRating: ${'⭐'.repeat(rating)}` : ''}`,
-        chatId: order.chatId,
-        buyerId: order.buyerId
-    }).catch(err => console.error('Confirm notification error:', err));
-
-    // Populate and return updated order
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, { order: updatedOrder }, "Delivery confirmed. Payment will be released by admin.")
-    );
-});
-
-// Buyer reports issue / requests return
-export const reportIssue = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const { reason, description, evidence, disputeVideoUrl } = req.body;
-    const buyerId = req.user._id;
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.buyerId.toString() !== buyerId.toString()) {
-        throw new ApiError(403, "Not authorized");
-    }
-
-    if (!reason) {
-        throw new ApiError(400, "Reason is required");
-    }
-
-    const validReasons = ['damaged_product', 'wrong_item', 'missing_item', 'not_as_described', 'defective', 'counterfeit', 'other'];
-    if (!validReasons.includes(reason)) {
-        throw new ApiError(400, `Invalid dispute reason. Must be one of: ${validReasons.join(', ')}`);
-    }
-
-    const reasonLabels = {
-        damaged_product: 'Damaged Product',
-        wrong_item: 'Wrong Item Received',
-        missing_item: 'Missing Item',
-        not_as_described: 'Not As Described',
-        defective: 'Defective Product',
-        counterfeit: 'Counterfeit Product',
-        other: 'Other'
-    };
-
-    order.orderStatus = 'disputed';
-    order.dispute = {
-        reason,
-        description,
-        evidence: evidence || [],
-        status: 'open',
-        createdAt: new Date()
-    };
-
-    if (disputeVideoUrl) {
-        order.dispute.disputeVideoUrl = disputeVideoUrl;
-        order.dispute.disputeVideoUploadedAt = new Date();
-    }
-
-    await order.save();
-
-    // Notify seller about the dispute
-    const reasonLabel = reasonLabels[reason] || reason;
-    sendOrderNotification({
-        recipientId: order.sellerId,
-        senderId: buyerId,
-        orderId: order._id,
-        orderNumber: order.orderNumber,
-        notificationMessage: `Dispute raised on order #${order.orderNumber}. Reason: ${reasonLabel}`,
-        chatMessageText: `A dispute has been raised on Order #${order.orderNumber}.\n\nReason: ${reasonLabel}${description ? `\nDescription: ${description}` : ''}${disputeVideoUrl ? '\n\nThe buyer has uploaded a proof video. You can view it in the order details.' : '\n\nThe buyer has not yet uploaded a proof video.'}\n\nThe admin team will review this dispute.`,
-        chatId: order.chatId,
-        buyerId: order.buyerId
-    }).catch(err => console.error('Dispute notification error:', err));
-
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, {
-            order: updatedOrder,
-            disputePolicy: {
-                videoRequired: true,
-                videoUploaded: !!disputeVideoUrl,
-                message: "You must upload a video showing proof of damage to be eligible for refund or return. This footage will be reviewed by both the admin and the seller. If valid proof is not provided, the item will not be eligible for refund or return."
-            }
-        }, "Issue reported, payment held until resolution")
-    );
-});
-
-// Buyer uploads dispute proof video (damage evidence)
-export const uploadDisputeVideo = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const { disputeVideoUrl } = req.body;
-    const buyerId = req.user._id;
-
-    if (!disputeVideoUrl) {
-        throw new ApiError(400, "Dispute video URL is required. You must upload a video showing proof of damage to be eligible for refund or return.");
-    }
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.buyerId.toString() !== buyerId.toString()) {
-        throw new ApiError(403, "Only the buyer can upload dispute video");
-    }
-
-    if (order.orderStatus !== 'disputed') {
-        throw new ApiError(400, "Can only upload dispute video for disputed orders");
-    }
-
-    if (!order.dispute) {
-        throw new ApiError(400, "No active dispute found for this order");
-    }
-
-    if (order.dispute.status === 'resolved' || order.dispute.status === 'rejected') {
-        throw new ApiError(400, "Cannot upload video for a closed dispute");
-    }
-
-    order.dispute.disputeVideoUrl = disputeVideoUrl;
-    order.dispute.disputeVideoUploadedAt = new Date();
-    await order.save();
-
-    // Notify seller that buyer has uploaded dispute proof video
-    sendOrderNotification({
-        recipientId: order.sellerId,
-        senderId: buyerId,
-        orderId: order._id,
-        orderNumber: order.orderNumber,
-        notificationMessage: `Buyer uploaded dispute proof video for order #${order.orderNumber}`,
-        chatMessageText: `The buyer has uploaded a proof video for the dispute on Order #${order.orderNumber}.\n\nYou can view the video evidence in the order details. The admin team will review this dispute.`,
-        chatId: order.chatId,
-        buyerId: order.buyerId
-    }).catch(err => console.error('Dispute video notification error:', err));
-
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, {
-            order: updatedOrder,
-            notice: "Your dispute video has been submitted. Both the admin and seller will review this footage. If valid proof of damage is not provided, the item will not be eligible for refund or return."
-        }, "Dispute proof video uploaded successfully")
-    );
-});
-
-// Buyer uploads payment proof
-export const uploadPaymentProof = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const { paymentScreenshot } = req.body;
-    const buyerId = req.user._id;
-
-    if (!paymentScreenshot) {
-        throw new ApiError(400, "Payment screenshot URL is required");
-    }
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.buyerId.toString() !== buyerId.toString()) {
-        throw new ApiError(403, "Only the buyer can upload payment proof");
-    }
-
-    // Validate order status
-    const allowedStatuses = ['payment_received', 'processing', 'shipped', 'delivered'];
-    if (!allowedStatuses.includes(order.orderStatus)) {
-        throw new ApiError(400, "Cannot upload payment proof for this order status");
-    }
-
-    order.buyerProof = {
-        ...order.buyerProof,
-        paymentScreenshot,
-        uploadedAt: new Date()
-    };
-    await order.save();
-
-    // Populate and return updated order
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, { order: updatedOrder }, "Payment proof uploaded successfully")
-    );
-});
-
-// Seller uploads packing video/images
-export const uploadPackingMedia = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const { packingVideoUrl, packingImages } = req.body;
-    const sellerId = req.user._id;
-
-    if (!packingVideoUrl && (!packingImages || packingImages.length === 0)) {
-        throw new ApiError(400, "At least one media file (video or images) is required");
-    }
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.sellerId.toString() !== sellerId.toString()) {
-        throw new ApiError(403, "Only the seller can upload packing media");
-    }
-
-    // Validate order status
-    const allowedStatuses = ['payment_received', 'processing', 'shipped'];
-    if (!allowedStatuses.includes(order.orderStatus)) {
-        throw new ApiError(400, "Cannot upload packing media for this order status");
-    }
-
-    // Initialize shippingInfo if it doesn't exist
-    if (!order.shippingInfo) {
-        order.shippingInfo = {};
-    }
-
-    // Update packing video if provided
-    if (packingVideoUrl) {
-        order.shippingInfo.packingVideoUrl = packingVideoUrl;
-    }
-
-    // Append new images to existing ones (limit to 10 total)
-    if (packingImages && packingImages.length > 0) {
-        const existingImages = order.shippingInfo.packingImages || [];
-        const allImages = [...existingImages, ...packingImages].slice(0, 10);
-        order.shippingInfo.packingImages = allImages;
-    }
-
-    await order.save();
-
-    // Populate and return updated order
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, { order: updatedOrder }, "Packing media uploaded successfully")
-    );
-});
-
-// Buyer uploads opening video
-export const uploadOpeningVideo = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const { openingVideoUrl } = req.body;
-    const buyerId = req.user._id;
-
-    if (!openingVideoUrl) {
-        throw new ApiError(400, "Opening video URL is required");
-    }
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.buyerId.toString() !== buyerId.toString()) {
-        throw new ApiError(403, "Only the buyer can upload opening video");
-    }
-
-    // Validate order status - only allow after shipment
-    const allowedStatuses = ['shipped', 'delivered'];
-    if (!allowedStatuses.includes(order.orderStatus)) {
-        throw new ApiError(400, "Can only upload opening video after shipment");
-    }
-
-    order.buyerProof = {
-        ...order.buyerProof,
-        openingVideoUrl,
-        uploadedAt: order.buyerProof?.uploadedAt || new Date()
-    };
-    await order.save();
-
-    // Populate and return updated order
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, { order: updatedOrder }, "Opening video uploaded successfully")
-    );
-});
-
-// Seller rates buyer
-export const rateBuyer = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const { rating, review } = req.body;
-    const sellerId = req.user._id;
-
-    if (!rating || rating < 1 || rating > 5) {
-        throw new ApiError(400, "Rating must be between 1 and 5");
-    }
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.sellerId.toString() !== sellerId.toString()) {
-        throw new ApiError(403, "Only the seller can rate the buyer");
-    }
-
-    if (!['confirmed', 'delivered'].includes(order.orderStatus)) {
-        throw new ApiError(400, "Order must be delivered or confirmed to rate the buyer");
-    }
-
-    if (order.sellerRating) {
-        throw new ApiError(400, "You have already rated this buyer");
-    }
-
-    order.sellerRating = rating;
-    order.sellerReview = review;
-    await order.save();
-
-    // Populate and return updated order
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, { order: updatedOrder }, "Buyer rated successfully")
-    );
-});
-
-// Buyer rates seller (dedicated endpoint, separate from confirmDelivery)
-export const rateSeller = asyncHandler(async (req, res) => {
-    const { orderId } = req.params;
-    const { rating, review } = req.body;
-    const buyerId = req.user._id;
-
-    if (!rating || rating < 1 || rating > 5) {
-        throw new ApiError(400, "Rating must be between 1 and 5");
-    }
-
-    const order = await Order.findById(orderId);
-    if (!order) {
-        throw new ApiError(404, "Order not found");
-    }
-
-    if (order.buyerId.toString() !== buyerId.toString()) {
-        throw new ApiError(403, "Only the buyer can rate the seller");
-    }
-
-    if (!['confirmed', 'delivered'].includes(order.orderStatus)) {
-        throw new ApiError(400, "Order must be delivered or confirmed to rate the seller");
-    }
-
-    if (order.buyerRating) {
-        throw new ApiError(400, "You have already rated this seller");
-    }
-
-    order.buyerRating = rating;
-    order.buyerReview = review;
-    await order.save();
-
-    // Populate and return updated order
-    const updatedOrder = await Order.findById(orderId)
-        .populate('buyerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('sellerId', 'fullName username profileImageUrl phoneNumber')
-        .populate('postId', 'media caption');
-
-    return res.status(200).json(
-        new ApiResponse(200, { order: updatedOrder }, "Seller rated successfully")
-    );
-});
-
-// Enhanced buyer order history with advanced filtering
-export const getBuyerOrderHistory = asyncHandler(async (req, res) => {
-    const {
-        status,
-        paymentStatus,
-        startDate,
-        endDate,
-        minAmount,
-        maxAmount,
-        search,
-        sortBy = 'createdAt',
-        sortOrder = 'desc',
-        page = 1,
-        limit = 20
-    } = req.query;
-
-    const buyerId = req.user._id;
-    const query = { buyerId };
-
-    // Status filters
-    if (status) query.orderStatus = status;
-    if (paymentStatus) query.paymentStatus = paymentStatus;
-
-    // Date range filter
-    if (startDate || endDate) {
-        query.createdAt = {};
-        if (startDate) query.createdAt.$gte = new Date(startDate);
-        if (endDate) query.createdAt.$lte = new Date(endDate);
-    }
-
-    // Amount range filter
-    if (minAmount || maxAmount) {
-        query.amount = {};
-        if (minAmount) query.amount.$gte = parseFloat(minAmount);
-        if (maxAmount) query.amount.$lte = parseFloat(maxAmount);
-    }
-
-    // Search by order number or product name
-    if (search) {
-        query.$or = [
-            { orderNumber: { $regex: search, $options: 'i' } },
-            { 'productDetails.name': { $regex: search, $options: 'i' } }
-        ];
-    }
-
-    // Sort configuration
-    const sortConfig = {};
-    sortConfig[sortBy] = sortOrder === 'asc' ? 1 : -1;
-
-    const orders = await Order.find(query)
-        .populate('sellerId', 'fullName username profileImageUrl')
-        .populate('postId', 'media caption')
-        .sort(sortConfig)
-        .skip((page - 1) * limit)
-        .limit(parseInt(limit));
-
-    const total = await Order.countDocuments(query);
-
-    // Calculate summary statistics
-    const stats = await Order.aggregate([
-        { $match: { buyerId: new mongoose.Types.ObjectId(buyerId) } },
-        {
-            $group: {
-                _id: null,
-                totalOrders: { $sum: 1 },
-                totalSpent: { $sum: '$amount' },
-                completedOrders: {
-                    $sum: { $cond: [{ $eq: ['$orderStatus', 'confirmed'] }, 1, 0] }
-                },
-                pendingOrders: {
-                    $sum: { $cond: [{ $in: ['$orderStatus', ['payment_received', 'processing', 'shipped', 'delivered']] }, 1, 0] }
-                },
-                cancelledOrders: {
-                    $sum: { $cond: [{ $in: ['$orderStatus', ['cancelled', 'refunded']] }, 1, 0] }
-                }
-            }
-        }
-    ]);
-
-    // Map 'held' → 'paid' for user display (admin retains 'held' in DB for manual release)
-    const userOrders = orders.map(o => {
-        const obj = o.toObject ? o.toObject() : o;
-        if (obj.paymentStatus === 'held') obj.paymentStatus = 'paid';
-        return obj;
-    });
-
-    return res.status(200).json(
-        new ApiResponse(200, {
-            orders: userOrders,
-            total,
-            page: parseInt(page),
-            totalPages: Math.ceil(total / limit),
-            stats: stats[0] || { totalOrders: 0, totalSpent: 0, completedOrders: 0, pendingOrders: 0, cancelledOrders: 0 }
-        }, "Buyer order history fetched")
-    );
-});
-
-// Enhanced seller order history with advanced filtering
-export const getSellerOrderHistory = asyncHandler(async (req, res) => {
-    const {
-        status,
-        paymentStatus,
-        startDate,
-        endDate,
-        minAmount,
-        maxAmount,
-        search,
-        sortBy = 'createdAt',
-        sortOrder = 'desc',
-        page = 1,
-        limit = 20
-    } = req.query;
-
-    const sellerId = req.user._id;
-    const query = { sellerId };
-
-    // Status filters
-    if (status) query.orderStatus = status;
-    if (paymentStatus) query.paymentStatus = paymentStatus;
-
-    // Date range filter
-    if (startDate || endDate) {
-        query.createdAt = {};
-        if (startDate) query.createdAt.$gte = new Date(startDate);
-        if (endDate) query.createdAt.$lte = new Date(endDate);
-    }
-
-    // Amount range filter
-    if (minAmount || maxAmount) {
-        query.amount = {};
-        if (minAmount) query.amount.$gte = parseFloat(minAmount);
-        if (maxAmount) query.amount.$lte = parseFloat(maxAmount);
-    }
-
-    // Search by order number or product name
-    if (search) {
-        query.$or = [
-            { orderNumber: { $regex: search, $options: 'i' } },
-            { 'productDetails.name': { $regex: search, $options: 'i' } }
-        ];
-    }
-
-    // Sort configuration
-    const sortConfig = {};
-    sortConfig[sortBy] = sortOrder === 'asc' ? 1 : -1;
-
-    const orders = await Order.find(query)
-        .populate('buyerId', 'fullName username profileImageUrl')
-        .populate('postId', 'media caption')
-        .sort(sortConfig)
-        .skip((page - 1) * limit)
-        .limit(parseInt(limit));
-
-    const total = await Order.countDocuments(query);
-
-    // Calculate summary statistics
-    const stats = await Order.aggregate([
-        { $match: { sellerId: new mongoose.Types.ObjectId(sellerId) } },
-        {
-            $group: {
-                _id: null,
-                totalOrders: { $sum: 1 },
-                totalEarned: { $sum: '$sellerAmount' },
-                completedOrders: {
-                    $sum: { $cond: [{ $eq: ['$orderStatus', 'confirmed'] }, 1, 0] }
-                },
-                pendingOrders: {
-                    $sum: { $cond: [{ $in: ['$orderStatus', ['payment_received', 'processing', 'shipped', 'delivered']] }, 1, 0] }
-                },
-                cancelledOrders: {
-                    $sum: { $cond: [{ $in: ['$orderStatus', ['cancelled', 'refunded']] }, 1, 0] }
-                }
-            }
-        }
-    ]);
-
-    // Map 'held' → 'paid' for user display (admin retains 'held' in DB for manual release)
-    const userOrders = orders.map(o => {
-        const obj = o.toObject ? o.toObject() : o;
-        if (obj.paymentStatus === 'held') obj.paymentStatus = 'paid';
-        return obj;
-    });
-
-    return res.status(200).json(
-        new ApiResponse(200, {
-            orders: userOrders,
-            total,
-            page: parseInt(page),
-            totalPages: Math.ceil(total / limit),
-            stats: stats[0] || { totalOrders: 0, totalEarned: 0, completedOrders: 0, pendingOrders: 0, cancelledOrders: 0 }
-        }, "Seller order history fetched")
-    );
-});
-
-// Get buyer order statistics
-export const getBuyerOrderStatistics = asyncHandler(async (req, res) => {
-    const buyerId = new mongoose.Types.ObjectId(req.user._id);
-    const { year, month } = req.query;
-
-    const matchQuery = { buyerId };
-
-    // Optional date filtering for specific year/month
-    if (year) {
-        const startDate = new Date(year, month ? month - 1 : 0, 1);
-        const endDate = month
-            ? new Date(year, month, 0, 23, 59, 59)
-            : new Date(year, 11, 31, 23, 59, 59);
-
-        matchQuery.createdAt = { $gte: startDate, $lte: endDate };
-    }
-
-    // Overall statistics
-    const overallStats = await Order.aggregate([
-        { $match: { buyerId: buyerId } },
-        {
-            $group: {
-                _id: null,
-                totalOrders: { $sum: 1 },
-                totalSpent: { $sum: '$amount' },
-                averageOrderValue: { $avg: '$amount' },
-                completedOrders: {
-                    $sum: { $cond: [{ $eq: ['$orderStatus', 'confirmed'] }, 1, 0] }
-                },
-                pendingOrders: {
-                    $sum: { $cond: [{ $in: ['$orderStatus', ['payment_received', 'processing', 'shipped', 'delivered']] }, 1, 0] }
-                },
-                disputedOrders: {
-                    $sum: { $cond: [{ $eq: ['$orderStatus', 'disputed'] }, 1, 0] }
-                },
-                cancelledOrders: {
-                    $sum: { $cond: [{ $in: ['$orderStatus', ['cancelled', 'refunded']] }, 1, 0] }
-                }
-            }
-        }
-    ]);
-
-    // Category-wise spending
-    const categoryStats = await Order.aggregate([
-        { $match: matchQuery },
-        {
-            $group: {
-                _id: '$productDetails.category',
-                totalSpent: { $sum: '$amount' },
-                orderCount: { $sum: 1 }
-            }
-        },
-        { $sort: { totalSpent: -1 } },
-        { $limit: 10 }
-    ]);
-
-    // Monthly spending trend (last 6 months)
-    const sixMonthsAgo = new Date();
-    sixMonthsAgo.setMonth(sixMonthsAgo.getMonth() - 6);
-
-    const monthlyTrend = await Order.aggregate([
-        {
-            $match: {
-                buyerId: buyerId,
-                createdAt: { $gte: sixMonthsAgo }
-            }
-        },
-        {
-            $group: {
-                _id: {
-                    year: { $year: '$createdAt' },
-                    month: { $month: '$createdAt' }
-                },
-                totalSpent: { $sum: '$amount' },
-                orderCount: { $sum: 1 }
-            }
-        },
-        { $sort: { '_id.year': 1, '_id.month': 1 } }
-    ]);
-
-    // Status breakdown
-    const statusBreakdown = await Order.aggregate([
-        { $match: { buyerId: buyerId } },
-        {
-            $group: {
-                _id: '$orderStatus',
-                count: { $sum: 1 }
-            }
-        }
-    ]);
-
-    return res.status(200).json(
-        new ApiResponse(200, {
-            overall: overallStats[0] || {},
-            categoryStats,
-            monthlyTrend,
-            statusBreakdown
-        }, "Buyer order statistics fetched")
-    );
-});
-
-// Get seller order statistics
-export const getSellerOrderStatistics = asyncHandler(async (req, res) => {
-    const sellerId = new mongoose.Types.ObjectId(req.user._id);
-    const { year, month } = req.query;
-
-    const matchQuery = { sellerId };
-
-    // Optional date filtering for specific year/month
-    if (year) {
-        const startDate = new Date(year, month ? month - 1 : 0, 1);
-        const endDate = month
-            ? new Date(year, month, 0, 23, 59, 59)
-            : new Date(year, 11, 31, 23, 59, 59);
-
-        matchQuery.createdAt = { $gte: startDate, $lte: endDate };
-    }
-
-    // Overall statistics
-    const overallStats = await Order.aggregate([
-        { $match: { sellerId: sellerId } },
-        {
-            $group: {
-                _id: null,
-                totalOrders: { $sum: 1 },
-                totalRevenue: { $sum: '$amount' },
-                totalEarned: { $sum: '$sellerAmount' },
-                platformFees: { $sum: '$platformFee' },
-                averageOrderValue: { $avg: '$amount' },
-                completedOrders: {
-                    $sum: { $cond: [{ $eq: ['$orderStatus', 'confirmed'] }, 1, 0] }
-                },
-                pendingOrders: {
-                    $sum: { $cond: [{ $in: ['$orderStatus', ['payment_received', 'processing', 'shipped', 'delivered']] }, 1, 0] }
-                },
-                disputedOrders: {
-                    $sum: { $cond: [{ $eq: ['$orderStatus', 'disputed'] }, 1, 0] }
-                },
-                cancelledOrders: {
-                    $sum: { $cond: [{ $in: ['$orderStatus', ['cancelled', 'refunded']] }, 1, 0] }
-                }
-            }
-        }
-    ]);
-
-    // Product-wise sales
-    const productStats = await Order.aggregate([
-        { $match: matchQuery },
-        {
-            $group: {
-                _id: '$productDetails.name',
-                totalRevenue: { $sum: '$amount' },
-                orderCount: { $sum: 1 },
-                averagePrice: { $avg: '$amount' }
-            }
-        },
-        { $sort: { orderCount: -1 } },
-        { $limit: 10 }
-    ]);
-
-    // Monthly revenue trend (last 6 months)
-    const sixMonthsAgo = new Date();
-    sixMonthsAgo.setMonth(sixMonthsAgo.getMonth() - 6);
-
-    const monthlyTrend = await Order.aggregate([
-        {
-            $match: {
-                sellerId: sellerId,
-                createdAt: { $gte: sixMonthsAgo }
-            }
-        },
-        {
-            $group: {
-                _id: {
-                    year: { $year: '$createdAt' },
-                    month: { $month: '$createdAt' }
-                },
-                totalRevenue: { $sum: '$amount' },
-                totalEarned: { $sum: '$sellerAmount' },
-                orderCount: { $sum: 1 }
-            }
-        },
-        { $sort: { '_id.year': 1, '_id.month': 1 } }
-    ]);
-
-    // Status breakdown
-    const statusBreakdown = await Order.aggregate([
-        { $match: { sellerId: sellerId } },
-        {
-            $group: {
-                _id: '$orderStatus',
-                count: { $sum: 1 }
-            }
-        }
-    ]);
-
-    // Average ratings
-    const ratingStats = await Order.aggregate([
-        {
-            $match: {
-                sellerId: sellerId,
-                buyerRating: { $exists: true, $ne: null }
-            }
-        },
-        {
-            $group: {
-                _id: null,
-                averageRating: { $avg: '$buyerRating' },
-                totalRatings: { $sum: 1 }
-            }
-        }
-    ]);
-
-    return res.status(200).json(
-        new ApiResponse(200, {
-            overall: overallStats[0] || {},
-            productStats,
-            monthlyTrend,
-            statusBreakdown,
-            ratings: ratingStats[0] || { averageRating: 0, totalRatings: 0 }
-        }, "Seller order statistics fetched")
-    );
-});
-
-// Export orders to CSV
-export const exportOrdersToCSV = asyncHandler(async (req, res) => {
-    const { type = 'buyer', status, startDate, endDate } = req.query;
-    const userId = req.user._id;
-
-    const query = type === 'buyer' ? { buyerId: userId } : { sellerId: userId };
-
-    if (status) query.orderStatus = status;
-    if (startDate || endDate) {
-        query.createdAt = {};
-        if (startDate) query.createdAt.$gte = new Date(startDate);
-        if (endDate) query.createdAt.$lte = new Date(endDate);
-    }
-
-    const orders = await Order.find(query)
-        .populate('buyerId', 'fullName username')
-        .populate('sellerId', 'fullName username')
-        .sort({ createdAt: -1 });
-
-    // Generate CSV content
-    const headers = [
-        'Order Number',
-        'Date',
-        'Product',
-        type === 'buyer' ? 'Seller' : 'Buyer',
-        'Amount',
-        'Order Status',
-        'Payment Status'
-    ].join(',');
-
-    const rows = orders.map(order => {
-        const counterParty = type === 'buyer'
-            ? (order.sellerId?.fullName || order.sellerId?.username || 'N/A')
-            : (order.buyerId?.fullName || order.buyerId?.username || order.buyerDetails?.fullName || 'Guest');
-
-        return [
-            order.orderNumber,
-            new Date(order.createdAt).toLocaleDateString(),
-            `"${order.productDetails.name}"`,
-            counterParty,
-            order.amount,
-            order.orderStatus,
-            order.paymentStatus
-        ].join(',');
-    });
-
-    const csv = [headers, ...rows].join('\n');
-
-    res.setHeader('Content-Type', 'text/csv');
-    res.setHeader('Content-Disposition', `attachment; filename="orders_${type}_${Date.now()}.csv"`);
-
-    return res.status(200).send(csv);
-});
-
-// Get all reviews for a user (as seller or as buyer)
-export const getUserReviews = asyncHandler(async (req, res) => {
-    const { userId } = req.params;
-    const { role = 'seller', page = 1, limit = 10 } = req.query;
-
-    const pageNum = parseInt(page);
-    const limitNum = parseInt(limit);
-    const userObjectId = new mongoose.Types.ObjectId(userId);
-
-    let matchQuery, ratingField, reviewField, reviewerPopulateField, reviewerRole;
-
-    if (role === 'buyer') {
-        // Reviews about this user as a buyer (ratings BY sellers)
-        matchQuery = {
-            buyerId: userObjectId,
-            sellerRating: { $exists: true, $ne: null }
-        };
-        ratingField = 'sellerRating';
-        reviewField = 'sellerReview';
-        reviewerPopulateField = 'sellerId';
-        reviewerRole = 'seller';
-    } else {
-        // Reviews about this user as a seller (ratings BY buyers)
-        matchQuery = {
-            sellerId: userObjectId,
-            buyerRating: { $exists: true, $ne: null }
-        };
-        ratingField = 'buyerRating';
-        reviewField = 'buyerReview';
-        reviewerPopulateField = 'buyerId';
-        reviewerRole = 'buyer';
-    }
-
-    // Aggregate stats
-    const statsResult = await Order.aggregate([
-        { $match: matchQuery },
-        {
-            $group: {
-                _id: null,
-                averageRating: { $avg: `$${ratingField}` },
-                totalReviews: { $sum: 1 },
-                fiveStars: { $sum: { $cond: [{ $eq: [`$${ratingField}`, 5] }, 1, 0] } },
-                fourStars: { $sum: { $cond: [{ $eq: [`$${ratingField}`, 4] }, 1, 0] } },
-                threeStars: { $sum: { $cond: [{ $eq: [`$${ratingField}`, 3] }, 1, 0] } },
-                twoStars: { $sum: { $cond: [{ $eq: [`$${ratingField}`, 2] }, 1, 0] } },
-                oneStar: { $sum: { $cond: [{ $eq: [`$${ratingField}`, 1] }, 1, 0] } }
-            }
-        }
-    ]);
-
-    const stats = statsResult[0] || {
-        averageRating: 0,
-        totalReviews: 0,
-        fiveStars: 0, fourStars: 0, threeStars: 0, twoStars: 0, oneStar: 0
-    };
-
-    // Paginated reviews
-    const reviews = await Order.find(matchQuery)
-        .select(`${ratingField} ${reviewField} createdAt productDetails.name orderNumber`)
-        .populate(reviewerPopulateField, 'fullName username profileImageUrl')
-        .sort({ createdAt: -1 })
-        .skip((pageNum - 1) * limitNum)
-        .limit(limitNum)
-        .lean();
-
-    // Normalize reviews into consistent shape
-    const normalizedReviews = reviews.map(order => ({
-        _id: order._id,
-        orderNumber: order.orderNumber,
-        rating: order[ratingField],
-        review: order[reviewField] || null,
-        productName: order.productDetails?.name || null,
-        reviewer: order[reviewerPopulateField],
-        reviewerRole,
-        createdAt: order.createdAt
-    }));
-
-    return res.status(200).json(
-        new ApiResponse(200, {
-            stats: {
-                averageRating: Math.round((stats.averageRating || 0) * 10) / 10,
-                totalReviews: stats.totalReviews,
-                breakdown: {
-                    5: stats.fiveStars,
-                    4: stats.fourStars,
-                    3: stats.threeStars,
-                    2: stats.twoStars,
-                    1: stats.oneStar
-                }
-            },
-            reviews: normalizedReviews,
-            page: pageNum,
-            totalPages: Math.ceil(stats.totalReviews / limitNum),
-            total: stats.totalReviews
-        }, "User reviews fetched successfully")
-    );
-});
+// Re-export shim — logic lives in ./order/
+export * from "./order/index.js";
